+++ src/iife.js
@@ -8,3 +8,4 @@ import './elementals/disclosure/index.js';
 import './elementals/menu/index.js';
 import './elementals/navbar/index.js';
 import './elementals/switch/index.js';
+import './elementals/tabs/index.js';