+++ src/index.js
@@ -6,3 +6,4 @@ export * from './elementals/disclosure/index.js';
 export * from './elementals/menu/index.js';
 export * from './elementals/navbar/index.js';
 export * from './elementals/switch/index.js';
+export * from './elementals/tabs/index.js';